fix(dotnet): alinha Score para double e Nivel para PascalCase em CriticidadeDto

+++ src/VestaNET.Application/DTOs/CriticidadeDto.cs
@@ -2,7 +2,7 @@ namespace VestaNET.Application.DTOs;
 
 public record CriticidadeDto(
     long IdAbrigo,
-    int Score,
+    double Score,
     string Nivel,
     string Justificativa,
     List<string> Recomendacoes);
